* p4a-packages: (49 commits)
  use _int instead of intptr_t for local stacks intptr_t is a #define on bsd
  accept changes due to bug fix in clean_unused_dynamic_variables also remove system dependant validation
  fix bugs in clean_unused_dynamic_variables formal parameters should not be cleaned
  validation paf_sipmabubble.c moved to LoopRecovering validation for statement insertion
  accept new result for loop recovering new version
  some more PRIdPTR instead of %ld
  Remove all warnings from pips 64 bits main issues where related to usage of int where intptr_t should be used also note the existence of the c99 standard macro PRIdPTR that expands to ld or d for printf output of intptr_t
  Make svn ignore Makefile.in generated by autoconf
  undo pips_config.h.in removal this file is important for taking into account result of pips configuration
  Upgrade loop_expansion to use statement insertion Loop expansion can be used either as a stand alone (using guards) or as a three steps process: 1/ call loop_expansion_init to create the expanded statement 2/ call statement_insertion to verify the validity of the insertion 3/ call loop_expansion to that will perform the expansion, guarded or unguarded depending on the result of statement_insertion A step further for terapix code generation (and terapix paper)
  upgrade statement insertion it is now based on out regions for dead code detection, and uses pragma extension to output its result so that further processings are made easier
  Upgrade loop detection in controlizer, pattern i=i-k was not checked as a possible increment
  -Improving the treatment of arrays : p = tab[i] => (p, tab[i][0],Exact) -Implementing points_to_general_assignment() which treats every kind of assignment. When it's nt a basic case the result is an ANYWHERE points-to relation. -Factorizing tests on lhs and rhs at the level of one function: points_to_assignment. -Changing approximation of heap allocated data structure from MAY to EXACT. -More basic cases added as pointer to data structure's field. -Add comments on the function's organization and on old/new functions. -Start treating points-to relations at declarations, not yet fully implemented...
  empty summation returns 0 instead of an error
  Extension to expression instructions for some Feautrier's example; to be extended later by Molka Becher
  test case for do loop recovery failure; Serge?
  Improved result for partial_eval02.c
  test case for Molka Becher; probably more test cases in fixpoint/EXEMPLES
  bug fix for Expressions/partial_eval02.c; to be checked better; comments added, reformatting,...
  In a previous commit, I fixed an error due to ECHO being defined twice this introduced a regression fixed here
  ...

+++ packages/PIPS/pips/src/Libs/effects-simple/prettyprint.c
@@ -42,9 +42,11 @@
 #include "properties.h"
 #include "linear.h"
 #include "ri.h"
+#include "effects.h"
 
 #include "misc.h"
 #include "ri-util.h"
+#include "effects-util.h"
 #include "text.h"
 
 #include "text-util.h"
@@ -114,8 +116,8 @@ compare_effect_reference(effect * e1, effect * e2)
 	if(unbounded_expression_p(e2))
 	  diff = -1;
 	else {
-	  int i1 = 0;
-	  int i2 = 0;
+	  intptr_t i1 = 0;
+	  intptr_t i2 = 0;
 
 	  /* FI: This is not enough as effects are not summarized
 	     right away. It may be impossible to find an integer value
